+++ section-2/step-06/src/main/java/com/carmanagement/agentic/tools/CleaningTool.java
@@ -26,7 +26,7 @@ public class CleaningTool {
      */
     @Tool("Requests a cleaning with the specified options")
     public String requestCleaning(
-            Long carNumber,
+            Integer carNumber,
             String carMake,
             String carModel,
             Integer carYear,
@@ -48,7 +48,7 @@ public class CleaningTool {
     }
 
     private String generateCleaningSummary(
-            Long carNumber,
+            Integer carNumber,
             String carMake,
             String carModel,
             Integer carYear,
